Windd version v0.15 This release brings improvements to several existing components: * Modified IsDeviceObject() to recognize an (optional) \\GLOBALROOT prefix   before Windows NT device namespace names * If possible, the function fd_reopen() now calculates the starting point   and range of an operation (`volume_offset' and `volume_length'), even in   case of a direct \\PhysicalDriveX access. * As an added bonus, these modifications also ensure that only a minimal   set of partitions gets locked (and dismounted) before each operation.

+++ compat/win32/diskmgmt/physical_disks.c
@@ -109,12 +109,12 @@ extern bool get_physdisk_path (TCHAR physdisk_path [])
   return (true);
 }
 
-extern long long get_physdisk_size (void)
+extern off_t get_physdisk_size (void)
 {
   if (not drives_physdisk_selected ())
-    return (0ll);
+    return ((off_t)0);
 
-  return ((long long)physical_disk[sel_disk].size_bytes);
+  return (physical_disk[sel_disk].size_bytes);
 }
 
 extern bool disk_contains_system_drive (void)
